Hey, so i took a stab at my humble attempt of trying to improve the documentation. I expect this to be an on-going effort, I hope this helps!
Let me know if you think this is helpful, I want to help this project succeed! I'm hoping if I pair my "making a game" with "learning to use this library", others will have an *easier* time integrating ecst with their game.

Anywho, love to talk this over with you.